Move instance credentials package to shared library
