+++ agent/ec2/ec2_client.go
@@ -16,7 +16,7 @@ package ec2
 import (
 	"strings"
 
-	"github.com/aws/amazon-ecs-agent/agent/credentials/instancecreds"
+	"github.com/aws/amazon-ecs-agent/ecs-agent/credentials/instancecreds"
 	"github.com/aws/amazon-ecs-agent/ecs-agent/ecs_client/model/ecs"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
