+++ ecs-agent/credentials/providers/credentials_filename_unsupported.go
