+++ ecs-agent/credentials/providers/credentials_filename_windows.go
