+++ ecs-agent/credentials/providers/rotating_shared_credentials_provider_test.go
